Extract AstIds into its own module
